evaldmon: Refactor Deserialization
During development it was noticed that BaseTeam components are not very
useful, as they contained BaseDrivers, making it very difficult to
convert driver list inside of the teams to Driver instances.

Therefore, instead of Base components, an DataEntry class was
introduced, which contains all deserialized data. A helper
DreamTeamComponentsCreator class was made to keep Controller content
low, and encapsulate component creation from controller.